Fix `update_cell` method conditions
Fix the `update_cell` method conditions that were misfiring due to the
mixes of Ranges and two matching conditions on `3` value for number of
neighbouring cells alive. Essentially split the conditions into alive
or dead cells, then look at the number of alive cells within the Ranges
expected, as according to Conway's Game of Life rules.